requirements.txt
python-nmap
:) enjoy!
